fix: change chat_id type to bigint

+++ init.sql
@@ -1,7 +1,7 @@
 CREATE TABLE aliases
 (
     id            SERIAL PRIMARY KEY,
-    chat_id       INT          NOT NULL,
+    chat_id       BIGINT       NOT NULL,
     sticker_id    VARCHAR(100) NOT NULL,
     sticker_alias VARCHAR(100) NOT NULL,
     UNIQUE (chat_id, sticker_alias)
